requirements.txt
fastapi
uvicorn[standard]
boto3
botocore
python-dotenv
python-jose
pydantic
slowapi
python-magic
requests